README with Decision Tree model
A decision tree is a supervised learning algorithm used for classification and regression modeling. Regression is a method used for predictive modeling, so these trees are used to either classify data or predict what will happen next. 


Decision trees look like flowcharts. They start at the root node with a specific question of data, which leads to branches that hold potential answers. The branches then lead to decision (internal) nodes, which ask more questions, leading to more outcomes. This continues until the data reaches a terminal (or “leaf”) node and ends.